MAINT: replace deprecated PixelDACopt with PixelDACoptFast
